requirements.txt
psycopg2-binary
pandas
numpy
ipykernel
hydra-core
omegaconf
tqdm